v2: *API Change* Lister.List() now returns []Resource, error
This change was made since a fsBundle could generate an error during
filepath.Walk, which was previously ignored. It's probably a better idea
to pass that along to the user. In v1 I chose to panic, I may change that
to this behaviour instead.

This was found using github.com/kisielk/errcheck