resources-go/dev: made development version branch
The main master version of the package will now be for unstable
development. API compatible backports of fixes should be possible.

This is currently incomplete (missing some url's and badges) and also
will be held in a separate branch for a grace period so that followers
can update their code.