Removed deadcode.
Due to the simplification of OpenPackageBundle, the checkPackage method is
unnecessary. This is because the function no longer loops, so it can only be
called from non-goroot packages anyway, so theres no need to check for that
eventuality. Also being called by the resources package should never happen
either unless intentionally as part of a test.

This should hopefully help with the flakyness of the tests.